+++ app/models/comment.py
@@ -12,10 +12,10 @@ class Comment(Base):
 
     # Project relations
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False, index=True)
-    user: Mapped["User"] = relationship("User", back_populates="comments", lazy="joined")  # noqa F821
+    user: Mapped["User"] = relationship("User", back_populates="comments", lazy="noload")  # noqa F821
 
     project_id: Mapped[int] = mapped_column(ForeignKey("projects.id"), nullable=False, index=True)
-    project: Mapped["Project"] = relationship("Project", back_populates="comments", lazy="joined")  # noqa F821
+    project: Mapped["Project"] = relationship("Project", back_populates="comments", lazy="noload")  # noqa F821
 
     def __repr__(self):
         return f"{self.__class__.__name__}(id={self.id}, project={self.project.name})"
