+++ app/tests/city/router/test_list.py
@@ -11,8 +11,8 @@ class TestCityList:
         assert response.status_code == 400
         assert response.json() == {"detail": "Токен отсутствует в заголовке"}
 
-    @pytest.mark.parametrize("num_requests, expected_rps", [(400, 230)])
-    async def test_rps(self, auth_ac_super, num_requests, expected_rps) -> None:
+    @pytest.mark.parametrize("num_requests, expected_rps, max_rps", [(400, 220, 250)])
+    async def test_rps(self, auth_ac_super, num_requests, expected_rps, max_rps) -> None:
         url = "/app/v1/cities/all"
         cookies = auth_ac_super.cookies.dict()
 
@@ -32,6 +32,9 @@ class TestCityList:
 
         # необязательная проверка минимального порога
         assert rps > expected_rps
+
+        # необязательная проверка максимального порога (чтобы видеть прирост, который не ожидал)
+        assert rps < max_rps
 
     async def test_list(self, auth_ac_super) -> None:
         response = await auth_ac_super.client.get("/app/v1/cities/all", cookies=auth_ac_super.cookies.dict())
